norm: execute
