+++ execute/internal/eval_utils.h
@@ -1,5 +1,5 @@
 #ifndef EVAL_UTILS_H
-#define EVAL_UTILS_H
+# define EVAL_UTILS_H
 
 // eval_utils.c
 void	new_pipeline(t_pipeline **pipeline);
@@ -10,7 +10,7 @@ void	new_argv(t_simple_command *sc);
 
 // new_redirect.c
 void	new_redirect_out(t_simple_command *sc,
-						 char *filename, t_node_type type);
+			 char *filename, t_node_type type);
 void	new_redirect_in(t_simple_command *sc, char *data, t_node_type type);
 
 #endif //EVAL_UTILS_H
