+++ execute/internal/execute_simple_command.c
@@ -1,5 +1,11 @@
+#include "execute_internal.h"
+#include "exit_status.h"
+#include "../../builtin/builtin.h"
+
 static void	execute_redirect(t_simple_command *sc);
-static bool	execute_builtin(t_executor *e, int argc, char **argv);
+static bool	execute_builtin(t_executor *e, int ac, char **av);
+static void	print_err_msg(char *path, char *msg);
+static void	handle_exec_error(char *path, bool is_exec_run);
 
 // execute_simple_command returns either child process pid
 // or macro 'CHILD_PROCESS_NOT_CREATED'
@@ -20,7 +26,7 @@ int	execute_simple_command(t_executor *e,
 	pid = x_fork();
 	if (pid == CHILD_PROCESS)
 	{
-		if (info->is_pipe && !info->is_last && info->pipefd[READ])
+		if (info->is_pipe && !info->is_last)
 			close(info->pipefd[READ]);
 		if (execute_builtin(e, sc->argc, sc->argv))
 			exit(EXIT_SUCCESS);
@@ -41,24 +47,59 @@ void	execute_redirect(t_simple_command *sc)
 		x_dup2(sc->r_out, STDOUT_FILENO);
 }
 
-bool	execute_builtin(t_executor *e, int argc, char **argv)
+bool	execute_builtin(t_executor *e, int ac, char **av)
 {
-	if (!ft_strcmp(argv[0], "cd"))
-		e->exit_status = builtin_cd(argc, argv, e->exit_status, e->env_vars);
-	else if (!ft_strcmp(argv[0], "pwd"))
-		e->exit_status = builtin_pwd(argc, argv, e->exit_status, e->env_vars);
-	else if (!ft_strcmp(argv[0], "exit"))
-		e->exit_status = builtin_exit(argc, argv, e->exit_status, e->env_vars);
-	else if (!ft_strcmp(argv[0], "echo"))
-		e->exit_status = builtin_echo(argc, argv, e->exit_status, e->env_vars);
-	else if (!ft_strcmp(argv[0], "export"))
-		e->exit_status = builtin_export(argc,
-										argv, e->exit_status, e->env_vars);
-	else if (!ft_strcmp(argv[0], "env"))
-		e->exit_status = builtin_env(argc, argv, e->exit_status, e->env_vars);
-	else if (!ft_strcmp(argv[0], "unset"))
-		e->exit_status = builtin_unset(argc, argv, e->exit_status, e->env_vars);
+	if (!ft_strcmp(av[0], "cd"))
+		e->exit_status = builtin_cd(ac, av, e->exit_status, e->env_vars);
+	else if (!ft_strcmp(av[0], "pwd"))
+		e->exit_status = builtin_pwd(ac, av, e->exit_status, e->env_vars);
+	else if (!ft_strcmp(av[0], "exit"))
+		e->exit_status = builtin_exit(ac, av, e->exit_status, e->env_vars);
+	else if (!ft_strcmp(av[0], "echo"))
+		e->exit_status = builtin_echo(ac, av, e->exit_status, e->env_vars);
+	else if (!ft_strcmp(av[0], "export"))
+		e->exit_status = builtin_export(ac, av, e->exit_status, e->env_vars);
+	else if (!ft_strcmp(av[0], "env"))
+		e->exit_status = builtin_env(ac, av, e->exit_status, e->env_vars);
+	else if (!ft_strcmp(av[0], "unset"))
+		e->exit_status = builtin_unset(ac, av, e->exit_status, e->env_vars);
 	else
 		return (false);
 	return (true);
+}
+
+void	print_err_msg(char *path, char *msg)
+{
+	ft_putstr_fd("minishell: ", STDERR_FILENO);
+	ft_putstr_fd(path, STDERR_FILENO);
+	ft_putstr_fd(": ", STDERR_FILENO);
+	ft_putendl_fd(msg, STDERR_FILENO);
+}
+
+void	handle_exec_error(char *path, bool is_exec_run)
+{
+	if (!is_exec_run)
+	{
+		print_err_msg(path, "command not found");
+		exit(ES_COMMAND_NOT_FOUND);
+	}
+	else
+	{
+		if (access(path, F_OK) == -1)
+		{
+			print_err_msg(path, "No such file or directory");
+			exit(ES_NO_SUCH_FILE);
+		}
+		else if (is_dir(path))
+		{
+			print_err_msg(path, "is a directory");
+			exit(ES_IS_A_DIRECTORY);
+		}
+		else if (access(path, X_OK) == -1)
+		{
+			print_err_msg(path, "Permission denied");
+			exit(ES_PERMISSION_DENIED);
+		}
+		exit(EXIT_SUCCESS);
+	}
 }
